pyproject.toml
[tool.poetry]
name = "connections-solver"
version = "0.1.0"
description = "NY Times Connections Game Solver"
authors = ["Sam Schneider <sschnei8@asu.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
selenium = "^4.21.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
duckdb = "^0.10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
